requirements.txt
APScheduler==3.6.3
astroid==2.12.5
autopep8==1.7.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
dill==0.3.5.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
pycodestyle==2.9.1
pylint==2.15.0
python-telegram-bot==13.13
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
wrapt==1.14.1
requests==2.32.4